+++ smart-dispatch-system/src/main/java/com/smartdispatch/websockets/NotificationService.java
@@ -3,8 +3,8 @@ package com.smartdispatch.websockets;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Service;
 
-import com.smartdispatch.dispatcher.domains.dtos.AssignmentDto;
-import com.smartdispatch.dispatcher.domains.dtos.IncidentDto;
+import com.smartdispatch.dispatcher.dtos.AssignmentDto;
+import com.smartdispatch.dispatcher.dtos.IncidentDto;
 import com.smartdispatch.websockets.websocketDto.*;
 
 @Service
@@ -39,7 +39,7 @@ public class NotificationService {
         notifyChannel("assignment/update", assignmentUpdateDto);
     }
 
-    public void notifyNewAssignment(Integer operatorId, AssignmentDto assignmentDto) {
+    public void notifyNewAssignment(Long operatorId, AssignmentDto assignmentDto) {
         notifyChannel("assignment/new/" + operatorId, assignmentDto);
     }
 }
